Module rich_sdl2_rust::audio::format

Type definitions for a format of audio data.

Structs

AudioFormat

A format to represent how is stored samples in audio data.

AudioFormatFlag

A flag to represent what type is used in audio data.