Struct rich_sdl2_rust::audio::MicrophoneDevice

pub struct MicrophoneDevice { /* private fields */ }

An audio device to input sound.

Implementations

impl MicrophoneDevice

pub fn all_devices() -> impl Iterator<Item = String>

Returns all of microphone audio device names on now.

pub fn open<'callback, T: AudioCallback<'callback>>( device: Option<String>, spec: &'callback AudioSpec<'callback, T>, fallback: FallbackFlag ) -> Result<(Self, AudioDeviceProperty)>

Opens the audio device named device with the specification and fallback flag. If device is None, the default audio device is used.

Errors

Returns Err if failed to open a specific microphone device.