Struct rich_sdl2_rust::audio::SpeakerDevice

pub struct SpeakerDevice { /* private fields */ }

An audio device to output sound.

Implementations

impl SpeakerDevice

pub fn all_devices() -> impl Iterator<Item = String>

Returns all of speaker audio device names on now.

pub fn open<'callback, T: AudioCallback<'callback>>(
device: Option<String>,
spec: &'callback AudioSpec<'callback, T>,
fallback: FallbackFlag
) -> Result<(Self, AudioDeviceProperty)>

Opens the audio device named device with the specification and fallback flag. If device is None, the default audio device is used.