Module rich_sdl2_rust::event

SDL2 event system and handlers.

Most event structures are defined in these sub modules, but crate::audio::event::AudioDeviceEvent is defined in the external module.

Modules

app

Events occurring on the application.

drop

Events on to drag and drop something to a window.

game_controller

A logical game controller that attached to a physical device.

joystick

A physical joystick device, controlling based on InputIndex and raw inputs.

keyboard

Events related on the keyboard.

mouse

Events and utilities for the mouse.

sensor

Setup and reading data from other sensors.

text

Controls of a simple text input field.

touch

Controls touch devices, fingers, and gestures.

window

Events for the window.

Structs

EventBox

It takes the closure of handler to register, and delivers events to your event handlers by polling.

Type Definitions

EventHandler

An event handler to receive some event structure,