Module rich_sdl2_rust::event::joystick

A physical joystick device, controlling based on InputIndex and raw inputs.

Modules

axis

Axes for a physical joystick.

button

Buttons for a physical joystick.

event

An event related on a joystick.

guid

The GUID of joystick devices.

hat

Hats for a physical joystick.

power_level

A representation of the power level of the battery in the joystick.

trackball

Trackballs for a physical joystick.

Structs

InputIndex

An index on the input device on the joystick. This is only generated from joystick input events from crate::EventBox.

Joystick

A physical joystick device manages its states, input devices and power levels.

JoystickId

An id of the joystick. It can be used to share/duplicate Joystick object, but not Send and Sync due to the safety.

JoystickSet

A set of all recognized Joystick at initialized.