Enum rich_sdl2_rust::event::joystick::event::JoystickEvent

pub enum JoystickEvent<'joystick> {
    Axis {
        timestamp: u32,
        id: JoystickId<'joystick>,
        axis: InputIndex,
        value: i16,
    },
    Trackball {
        timestamp: u32,
        id: JoystickId<'joystick>,
        trackball: InputIndex,
        x_amount: i16,
        y_amount: i16,
    },
    Button {
        timestamp: u32,
        id: JoystickId<'joystick>,
        button: InputIndex,
        is_pressed: bool,
    },
    Hat {
        timestamp: u32,
        id: JoystickId<'joystick>,
        hat: InputIndex,
        value: PovHat,
    },
    DeviceAdded {
        timestamp: u32,
        joystick: Joystick,
    },
    DeviceRemoved {
        timestamp: u32,
        id: JoystickId<'joystick>,
    },
}

An event occurs on inputted from a joystick or changed a joystick.

Variants

Axis

An Axis was changed,

Fields of Axis

timestamp: u32

When this event occurred.

id: JoystickId<'joystick>

The id of a joystick that has an axis.

axis: InputIndex

The index of an axis.

value: i16

The changed value.

Trackball

A trackball was changed,

Fields of Trackball

timestamp: u32

When this event occurred.

id: JoystickId<'joystick>

The id of a joystick that has a trackball.

trackball: InputIndex

The index of a trackball.

x_amount: i16

The changed value, scroll amounts of x direction.

y_amount: i16

The changed value, scroll amounts of y direction.

Button

A button was changed.

Fields of Button

timestamp: u32

When this event occurred.

id: JoystickId<'joystick>

The id of a joystick that has a button.

button: InputIndex

The index of a button.

is_pressed: bool

Whether the button was pressed.

Hat

A hat was changed.

Fields of Hat

timestamp: u32

When this event occurred.

id: JoystickId<'joystick>

The id of a joystick that has a hat.

hat: InputIndex

The index of a hat.

value: PovHat

The changed value.

DeviceAdded

A joystick was added.

Fields of DeviceAdded

timestamp: u32

When this event occurred.

joystick: Joystick

An added joystick.

DeviceRemoved

The joystick was removed.

Fields of DeviceRemoved

timestamp: u32

When this event occurred.

id: JoystickId<'joystick>

The id of the removed joystick.