Module rich_sdl2_rust::haptic

Haptic devices, which give the players to feedback by some force.

Modules

direction
Defining directions and coordinate systems of the haptic movements.

effect
Controls the effect on the haptic device.

rumble
Rumbling the haptic device.

Structs

Haptic
A haptic device.

HapticProperty
A property of the haptic device.

HapticSet
All of recognized haptic devices at initialized.

JoystickHaptic
A haptic device got from the joystick.

MouseHaptic
A haptic device got from the mouse.

PausedHaptic
A haptic device but frozen not to interact.

PendingEffect
An haptic effect but pending to play.

PlayingEffect
A playing haptic effect.