Struct rich_sdl2_rust::haptic::Haptic

pub struct Haptic { /* private fields */ }

A haptic device.

Implementations

impl Haptic

pub fn name(&self) -> String

Returns the name of the haptic device.

pub fn num_axes(&self) -> u32

Returns the numbers of the axes on the haptic device.

pub fn is_effect_supported(&self, effect: &HapticEffect) -> bool

Returns whether the effect is supported on the haptic device.

pub fn new_effect(&self, effect: &HapticEffect) -> Result<PendingEffect<'_>>

Constructs the PendingEffect from the effect specification.

Errors

Returns Err if this feature is unsupported, or failed to create a new haptic effect on the device.

pub fn effects_creation_capacity(&self) -> usize

Returns the capacity of the effects on the haptic device.

pub fn effects_playing_capacity(&self) -> usize

Returns the maximum numbers of playing the effects at same time on the haptic device.

pub fn stop_all_effect(&self)

Stops all the playing effect.

pub fn set_gain(&self, gain: u32)

Sets the global gain. If not supported, this has no effects.

pub fn set_auto_center(&self, auto_center: u32)

Sets auto-center. If not supported, this has no effects.

pub fn property(&self) -> HapticProperty

Queries a property on the haptic device.

pub fn pause(self) -> PausedHaptic

Pauses the haptic device and converts into PausedHaptic.