Module rich_sdl2_rust::hint

Getting/setting hints for SDL2 features working.

Structs

HintEvent

An event on updated the hint.

HintObserver

An hint update observer.

Enums

HintPriority

A priority of the hint specifying.

Traits

HintCallback

A callback invoked on updated the hint.

Functions

clear_hints

Clears all the set hints.

get_hint

Returns the hint value of the key if exists.

get_hint_bool

Returns the boolean hint value of the key if exists.

set_hint

Sets the hint value of the key, or Err if does not exist.

set_hint_with_priority

Sets a hint value of the key with a priority.