Module rich_sdl2_rust::mixer::chunk::channel

Channels to play a MixChunk.

Modules

effect

Effect processing for the channel, such as panning, decaying, etc.

Structs

Channel

A mixing channel for playing a MixChunk.

ChannelGroup

A channel group, reserving some Channels.

Pauser

A pauser to pause playing audio on the Channel. Unpause to drop this.

PlayOptions

Options to play a chunk by Channel::play.

Enums

PlayLoops

Loops on playing in PlayOptions.