Struct rich_sdl2_rust::mixer::chunk::channel::Channel

pub struct Channel<'device>(_, _);

A mixing channel for playing a MixChunk.

Implementations

impl<'device> Channel<'device>

pub fn volume(&self) -> u32

Returns the output volume of the channel. The volume is in 0..=128.

pub fn set_volume(&self, volume: u32)

Sets the output volume of the channel. The volume is clamped to 0..=128.

pub fn play(self, chunk: &MixChunk<'_>, options: PlayOptions) -> Result<Self>

Starts to play a chunk.

pub fn play_fade_in(
self,
chunk: &MixChunk<'_>,
fade_in: u32,
options: PlayOptions
) -> Result<Self>

Starts to play a chunk with fade-in time in milliseconds.

pub fn pause(&'device mut self) -> Option<Pauser<'device>>

Pauses playing and returns the Pauser, or None if it is free.

pub fn halt(&self)

Halts playing on the channel.

pub fn fade_out(&self, fade_out: u32) -> usize

Fade out playing on the channel in milliseconds. And returns the numbers of channels that is fading out.

pub fn is_playing(&self) -> bool

Returns whether the channel is playing.

pub fn playing_chunk(&self) -> Option<MixChunk<'_>>

Returns the playing chunk if exists.