Module rich_sdl2_rust::renderer

Renderer for a window, to render some geometries or query driver information.

Modules

clip

Clipping a renderer.

driver

Drivers of renderers.

info

Defining information structures of a renderer.

pen

Rendering geometries and setting colors.

Structs

PasteExFlip

Flip mode on pasting from another texture.

PasteExOption

An option for PasteExt::paste_ex.

Renderer

A SDL2 renderer. This is often used for rendering with pen::Pen.

Traits

PasteExt

An extension for Renderer to paste from another texture.