Struct rich_sdl2_rust::texture::Texture

pub struct Texture<'renderer> { /* private fields */ }

A texture used to draw the image to Renderer.

Implementations

impl<'renderer> Texture<'renderer>

pub fn new(
renderer: &'renderer Renderer<'renderer>,
access: TextureAccess
) -> Result<Self>

Constructs a texture from the renderer with access type.

Errors

Returns Err if failed to allocate, no rendering context was active, the format was unsupported, or the width or height were out of range.

pub fn from_surface(
renderer: &'renderer Renderer<'renderer>,
surface: &'renderer impl Surface
) -> Self

Constructs a texture from the Surface. The texture will be readonly and the access type will be TextureAccess::Static.

pub fn alpha_mod(&self) -> u8

Returns the alpha mod of the texture.

pub fn set_alpha_mod(&self, alpha: u8) -> Result<()>

Sets the alpha mod of the texture.

Errors

Returns Err if setting the alpha mod is unsupported.

Panics

Panics if some unrecoverable error is occurred.

pub fn color_mod(&self) -> Rgb

Returns the color mod of the texture.

pub fn set_color_mod(&self, _: Rgb)

Sets the color mod of the texture.

pub fn lock(&'renderer mut self, area: Option<Rect>) -> Lock<'renderer>

Obtains the lock for the texture in area, or whole if None.

pub fn clip(&self) -> &Option<Rect>

Return the clip area of the texture if available.

pub fn set_clip(&mut self, clip: Option<Rect>)

Sets the clip area.

pub fn bind_to_current_gl_context(&self) -> Result<(f32, f32)>

Binds the texture to the current OpenGL context. And returns the size in the context.

Errors

Returns Err if failed to bind the texture.

pub fn unbind_from_current_gl_context(&self) -> Result<()>

Unbinds the texture to the current OpenGL context.

Errors

Returns Err if failed to unbind the texture.