Module rich_sdl2_rust::ttf::font

Font data managers.

Modules

glyph

Glyph metrics.

pen

Extensions for crate::renderer::pen::Pen.

Structs

Dpi

A pixel density per inch.

Font

A font data structure.

KerningDisabler

A disabler of font kerning. Kerning of the font is disabled until this struct is dropped.

TtfSurface

A surface of TrueType Font.

Enums

FontHinting

A hinting for a font.

FontStyle

A text style of a font.

RenderMode

A render mode for the font glyphs.

Traits

AttributeExt

An extension for getting the attributes of the fonts.

FontSetting

An extension for settings of the font.

MetricExt

An extension for getters of the font metrics.

RenderExt

An extension for rendering the text.

StyleExt

An extension of FontStyle and outline width getters/setters.