Struct rich_sdl2_rust::window::Window

pub struct Window<'video> { /* private fields */ }

A window made by SDL2.

Implementations

impl<'video> Window<'video>

pub fn from_id(id: u32, video: &'video Video<'_>) -> Option<Self>

Gets a window from the window id, or None if does not exist.

pub fn grabbed(video: &'video Video<'_>) -> Option<Self>

Gets a grabbed window, or None if does not exist.

pub fn mouse_focused(video: &'video Video<'_>) -> Option<Self>

Gets a focused window, or None if does not exist.

pub fn state(&self) -> WindowState

Returns the state of the window.

pub fn display(&self) -> Option<Display<'_>>

Returns the display at the window, or None if unavailable.

pub fn id(&self) -> u32

Returns the window id.

pub fn pixel_format(&self) -> PixelFormatKind

Returns the pixel format of the window context.

pub fn show(&self)

Shows the window.

pub fn hide(&self)

Hides the window.

pub fn raise(&self)

Raises the window.

pub fn full_screen(&self) -> Result<()>

Make the window full screen.

Errors

Returns Err if failed to make it full screen.

pub fn maximize(&self)

Maximizes the window.

pub fn minimize(&self)

Minimizes the window.

pub fn restore(&self)

Restores the window from maximization/minimization.

pub fn set_icon(&self, icon: &impl Surface)

Sets an icon from a surface for the window.

pub fn is_screen_keyboard_shown(&self) -> bool

Returns whether the window is showing the screen keyboard.

pub fn surface(&self) -> WindowSurface<'_>

Makes the window surface.