Module rich_sdl2_rust::audio

Provides audio device control, configuration, wav format utilities and so on.

Modules

buffer

An audio buffer AudioBuffer<T> with a format, sample rates, numbers of channels and a buffer. It can be convert into another format and mix by the specified volume.

event

An event related on an audio device.

format

Type definitions for a format of audio data.

queue

Structures to use an audio device by pushing methods.

spec

Provides tools to make a specification to require what an audio device is.

status

A status of an audio device.

stream

An audio stream allows to read and write by streaming method.

wav

A structure for wav format.

Structs

AudioDeviceLock

A lock to prevent an audio device from calling the callback spec::AudioCallback.

AudioDeviceProperty

A property of an audio device.

MicrophoneDevice

An audio device to input sound.

SpeakerDevice

An audio device to output sound.

Traits

AudioDevice

Common methods for an audio device, such as a speaker and a microphone.

Functions

all_audio_drivers

Returns all of audio driver names recognized on now.

current_driver

Returns the current audio driver name, or None if it does not exist.

device_name

Returns the device name of the id device_id. Please set is_microphone according to what type you want to.