Struct rich_sdl2_rust::event::EventBox

pub struct EventBox<'video> { /* private fields */ }

It takes the closure of handler to register, and delivers events to your event handlers by polling.

Implementations

impl<'video> EventBox<'video>

pub fn new(_: &'video Video<'_>) -> Self

Constructs an event box from the video system.

pub fn handle_quit(&mut self, handler: EventHandler<'video, QuitEvent>)

Registers the handler to handle QuitEvent.

pub fn handle_window(&mut self, handler: EventHandler<'video, WindowEvent>)

Registers the handler to handle WindowEvent.

pub fn handle_keyboard(&mut self, handler: EventHandler<'video, KeyboardEvent>)

Registers the handler to handle KeyboardEvent.

pub fn handle_input(&mut self, handler: EventHandler<'video, TextInputEvent>)

Registers the handler to handle TextInputEvent.

pub fn handle_editing( &mut self, handler: EventHandler<'video, TextEditingEvent> )

Registers the handler to handle TextEditingEvent.

pub fn handle_mouse(&mut self, handler: EventHandler<'video, MouseEvent>)

Registers the handler to handle MouseEvent.

pub fn handle_controller( &mut self, handler: EventHandler<'video, ControllerEvent<'video>> )

Registers the handler to handle ControllerEvent.

pub fn handle_joystick( &mut self, handler: EventHandler<'video, JoystickEvent<'video>> )

Registers the handler to handle JoystickEvent.

pub fn handle_audio_device( &mut self, handler: EventHandler<'video, AudioDeviceEvent> )

Registers the handler to handle AudioDeviceEvent.

pub fn handle_drop(&mut self, handler: EventHandler<'video, DropEvent>)

Registers the handler to handle DropEvent.

pub fn handle_gesture(&mut self, handler: EventHandler<'video, GestureEvent>)

Registers the handler to handle GestureEvent.

pub fn poll(&self)

Polling the events and triggers the event handlers.

pub fn wait_next_event_with(&self, timeout_ms: u32)

Waits until the next event occurs, but unlock with timeout seconds.