Struct rich_sdl2_rust::file::RwOps

pub struct RwOps<'a> { /* fields omitted */ }

A file handler, how to read and write from file on SDL2.

Implementations

impl<'a> RwOps<'a>

pub unsafe fn ptr(&self) -> NonNull<SDL_RWops>

Returns the raw pointer of the file handler.

Safety

Dereferencing the pointer and using the field is not recommended. Please use carefully.

pub fn from_file(file_name: &str, mode: OpenMode) -> Result<RwOps<'static>>

Constructs from file name with the open mode, or Err on failure.

pub fn from_mem(buf: &'a [u8]) -> Result<Self>

Constructs from the buffer &[u8], or Err on failure.

pub fn from_mem_mut(buf: &'a mut [u8]) -> Result<Self>

Constructs from the mutable buffer &mut [u8], or Err on failure.

pub fn size(&self) -> Result<usize>

Returns the size of the read/write target.

pub fn tell(&mut self) -> Result<u64>

Returns the current position of seeking.

pub fn read_u8(&mut self) -> u8

Reads and pops the 8-bits value.

pub fn read_be16(&mut self) -> u16

Reads and pops the big endian 16-bits value.

pub fn read_le16(&mut self) -> u16

Reads and pops the little endian 16-bits value.

pub fn read_be32(&mut self) -> u32

Reads and pops the big endian 32-bits value.

pub fn read_le32(&mut self) -> u32

Reads and pops the little endian 32-bits value.

pub fn read_be64(&mut self) -> u64

Reads and pops the big endian 64-bits value.

pub fn read_le64(&mut self) -> u64

Reads and pops the little endian 64-bits value.

pub fn write_u8(&mut self, value: u8) -> bool

Writes the 8-bits value.

pub fn write_be16(&mut self, value: u16) -> bool

Writes the big endian 16-bits value.

pub fn write_le16(&mut self, value: u16) -> bool

Writes the little endian 16-bits value.

pub fn write_be32(&mut self, value: u32) -> bool

Writes the big endian 32-bits value.

pub fn write_le32(&mut self, value: u32) -> bool

Writes the little endian 32-bits value.

pub fn write_be64(&mut self, value: u64) -> bool

Writes the big endian 64-bits value.

pub fn write_le64(&mut self, value: u64) -> bool

Writes the little endian 64-bits value.