Module rich_sdl2_rust::haptic::effect

Controls the effect on the haptic device.

Structs

Condition

A condition to trigger the effect. Refer to Direction for which side is the positive/negative on the joystick.

Custom

A custom periodic waveform by your sampled data.

Envelope

An envelope to fade in/out the effect. If both attack_length and fade_level are 0, the envelope is not used. An example of a constant effect evolution in time:

Level

A magnitude level of the force.

Play

Length and delay of the playing.

Ramp

A linear ramp to interpolate the force of the effect.

Trigger

A trigger button to start the effect, and an interval between the effect.

Vector3

A vector to represent the XYZ component for Condition.

Wave

A periodic waveform specification.

Enums

HapticEffect

An effect on the haptic device.

WaveKind

A kind of the waveform.