Struct rich_sdl2_rust::renderer::Renderer

pub struct Renderer<'window> { /* fields omitted */ }

A SDL2 renderer. This is often used for rendering with pen::Pen.

Implementations

impl<'window> Renderer<'window>

pub fn new(window: Window<'window>) -> Self

Constructs a renderer from the window.

pub fn clip(&'window mut self, area: Rect) -> ClippedRenderer<'window>

Clips the renderer by area.

pub fn set_target<'texture: 'window>(
&'window self,
texture: &'texture Texture<'_>
)

Sets the render target to the texture.

impl Renderer<'_>

pub fn window(&self) -> &Window<'_>

Returns the reference of Window.

pub fn output_size(&self) -> Size

Returns the geometry size of the output from the renderer.

pub fn logical_size(&self) -> Option<Size>

Returns the logical size of the renderer if available.

pub fn set_logical_size(&self, _: Size)

Sets the logical size of the renderer.

pub fn is_forced_integer_scale(&self) -> bool

Returns whether integer scaled is forced.

pub fn force_integer_scale(&self, enabled: bool)

Sets whether integer scaled is forced.

pub fn scale(&self) -> Scale

Returns the scale of th renderer.

pub fn set_scale(&self, _: Scale)

Sets the scale of the renderer.

pub fn viewport(&self) -> Rect

Returns the viewport rectangle of the renderer.

pub fn set_viewport(&self, area: Option<Rect>)

Sets the viewport rectangle of the renderer.

pub fn set_target_default(&self)

Resets the render target to the original window.