Trait rich_sdl2_rust::surface::bmp::BmpSaveExt

pub trait BmpSaveExt {
    fn save_bmp(&self, file_name: &str) -> Result<(), BmpSaveError>;
}

An extension for a Surface to save the image as BMP format.

Required Methods

fn save_bmp(&self, file_name: &str) -> Result<(), BmpSaveError>

Saves the surface image as BMP format.

Errors

Returns Err if failed to save an image to the file.