Struct rich_sdl2_rust::ttf::font::Font

pub struct Font<'ttf> { /* private fields */ }

A font data structure.

Implementations

impl<'ttf> Font<'ttf>

pub fn new( _ttf: &'ttf Ttf, file_name: &str, point: u32, index: Option<usize> ) -> Result<Self>

Constructs a font data from file name and point size. The font face in the font file can be selected by an optional index, which will be 0 if None.

Panics

Panics if file_name is empty.

pub fn glyph(&self, ch: char) -> Option<Glyph<'_>>

Returns the glyph of the font if exists.

pub fn rendered_size(&self, text: &str) -> Result<Size>

Calculates the rendered size of the text, or Err on failure. The height will be same as MetricExt::height.

pub fn rendered_width( &self, text: &str, expected_width: u32 ) -> Result<(u32, usize)>

Returns the rendered width in pixels and the numbers of characters until reaching expected_width. Err will be returned on out of memory.