Module rich_sdl2_rust::window

Window managements, graphics and mouse controls.

Modules

cursor

Cursor control on a window.

message_box

Modal control for a window.

Structs

BorderWidths

A border widths for the window.

Brightness

A brightness in the window.

Gamma

A gamma ramps for a window.

GammaParam

A gamma ramps for setting to a window.

HitTest

A hit tester for a window.

Opacity

An opacity of the window.

Position

A position of the window.

Window

A window made by SDL2.

WindowBuilder

A builder for the Window.

WindowCoord

A coordinate value for the window position.

WindowState

A state of a super::Window.

Enums

HitTestResult

A hit test result that determined cursor from the mouse position.

WindowContextKind

A kind of a window’s context.

WindowFormat

A format of a super::Window.

Traits

BorderExt

An extension for Window to query border widths.

BrightnessExt

An extension for Window to get/set the brightness.

ConfigExt

An extension for Window to configure its properties.

GammaExt

An extension for Window to get/set the gamma ramp.

HitTester

A callback to hit test with the mouse position.

MouseGrabExt

An extension for Window to grab/ungrab the mouse.